Remove address collection from create_payment call
Address from the pop up is used to create a new address
and save it as the payment_source(Spree::CreditCard) address.
Since, we don't collect this address, we remove this process.
Spree::CreditCard will have address as nil but it shouldn't have
any implications. Since order.bill_address can be considered as
the billing address